ValueError: Must pass 2-d input. shape=(395216, 5, 1) --> Fix
The error message indicates that you are trying to pass a 3-dimensional NumPy array to pd.DataFrame() which expects a 2-dimensional array. The shape of the input array is (395216, 5, 1) which means it has 395216 rows, 5 columns, and 1 channel.

To fix this error, you need to reshape the array to 2 dimensions by flattening the last dimension.